Rename training example scripts
